+++ src/data/entities/points.rs
@@ -6,8 +6,8 @@ use eframe::egui::plot::PlotPoint;
 #[sea_orm(table_name = "points")]
 pub struct Model {
 	#[sea_orm(primary_key, auto_increment = false)]
-	pub id: i32,
-	pub metric_id: i32,
+	pub id: i64,
+	pub metric_id: i64,
 	pub x: f64,
 	pub y: f64,
 }
